I build my own timer for the planner.

Sometimes i get similar results as @Sebastianv650, but mostly total different.
The values include the time used by the stepper interrupt.
With the same model, depending on the feedrate i can get maximum times from 4 to 20ms and averages from 2 to 8ms. Without any ADVANCE.

In a second step i minimized the time for the stepper-ISR by short cutting it.
Then, for the same model i get now max at 4ms and average at 1.73ms for the planing time + the remaining interrupts.

Sadly non of this value can be used for `MIN_BLOCK_TIME` the variation is too big.

(cc @thinkyhead, @Sebastianv650)