@Sebastianv650 2016-12-10
Personaly I would use micros instead of millis. For the 20ms region it will not change anything, but an average of 1.73ms where each result has a resolution of 1ms is critical in my eyes.

It's true, the runtime can vary a hughe amount. That's why I was measuring only XYE moves and only displaying the average time over 100 calculations, combined with one gcode section as a "benchmark".
The 6ms min. block time was working for me - but I have only one printer to test the reason for it was `SLOWDOWN` wasn't working for me due to the long time a update of the LCD took. It's on my list to tests with the new RCBugFix and slowdown again this weekend. If it runs my short moves test gcode without a buffer underrun, maybe we can get rid of the ensure smooth moves thing.
Again, I have only one cartesian printer with one specific display - sadly I can't say how it perfoms on other configs.