@AnHardt 2016-12-10
Having seeped abut this, i see, timing the planner with the shortcut in the stepper-ISR is crap.
Having always 0 planed moves in the buffer prevents from recalculating already planed moves - so alters the result considerably. :-(

I really like the idea of stopping the display updates when we are in danger the buffer is running dry.
But for finding a good criteria when this is the case, i have no good idea.
Ether is shuting nown the updates much to early or not not early enough.