@Sebastianv650 2016-12-11
I think there is no perfect solution. Marlin doesn't know what will come next with it's small buffer.
Maybe it would work better for a printer with a random configuration and display to keep track of the update runtime on it's own? The display update could track the maximum needed time for an update, starting from 0 on a new print. Then, long_move() could compare the theoretical runtime of the buffer (which will always be longer in reality due to acceleration and deceleration) to the calculated max. runtime.
So it could eleminate the need for a given value in the config.

The second parameter with min. segment time can be omitted if slowdown does a good job. I will try testing slowdown today if it can prevent a buffer underrun in all situations.