@Sebastianv650 2016-12-11
Maybe I'm doing something wrong? No mater what I'm trying, I can't get the buffer to 0 with the new LCD updates.
I'm using a modified ADVANCED_OK to display the buffer status. Instead of displaying the free space, I changed it to display the used space because that's easier to see a low number. I also tested an alternative where I set a flag if `current_block = planner.get_current_block();` isn't able to get a new one.

Even with `SLOWDOWN `disabled I don't reach 0. I have no idea why, but I take it as a good sign...