@AnHardt 2016-12-13
Replaced by https://github.com/MarlinFirmware/Marlin/pull/5495